requirements.txt
Django==4.2.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
django-cors-headers==4.3.1
django-extensions==3.2.3

pymysql==1.1.1
sqlparse==0.5.3
python-decouple==3.8
pillow==9.5.0
zeep==4.3.1

django-jalali==6.0.1
jdatetime==5.0.0
az-iranian-bank-gateways>=2.1,<3
